Resolve "create helper to support new users and developers to dive into"

Closes #8

See merge request rapiddweller/benerator/rapiddweller-benerator-ce!3

+++ src/test/java/com/rapiddweller/benerator/test/ValidatorMock.java
@@ -57,7 +57,7 @@ public class ValidatorMock implements Validator<Integer> {
 
 	@Override
 	public boolean valid(Integer candidate) {
-		return (candidate != null && value == candidate.intValue());
+		return (candidate != null && value == candidate);
 	}
 
 }
